refactor: ItemException 계층화 및 ErrorStatus 삭제
