+++ README.md
@@ -108,5 +108,3 @@ http://localhost:8080/items/1
 ## 상품 요구사항
   - [x] 상품의 이름은 최소 1자, 최대 50자까지 가능하다.
   - [x] 상품의 금액은 최소 10원, 최대 1억원 까지 가능하다.
-
-
