+++ src/main/java/cart/exception/item/ItemFieldNotValidException.java
@@ -1,0 +1,10 @@
+package cart.exception.item;
+
+import org.springframework.http.HttpStatus;
+
+public class ItemFieldNotValidException extends ItemException {
+
+    public ItemFieldNotValidException(String message) {
+        super(HttpStatus.BAD_REQUEST, message);
+    }
+}
