+++ src/main/java/cart/model/ItemPrice.java
@@ -1,7 +1,6 @@
 package cart.model;
 
-import cart.exception.ErrorStatus;
-import cart.exception.ItemException;
+import cart.exception.item.ItemFieldNotValidException;
 import java.math.BigInteger;
 
 public class ItemPrice {
@@ -19,7 +18,7 @@ public class ItemPrice {
 
     private void validatePrice(BigInteger price) {
         if (MIN_PRICE.compareTo(price) > 0 || MAX_PRICE.compareTo(price) < 0) {
-            throw new ItemException(ErrorStatus.PRICE_RANGE_ERROR);
+            throw new ItemFieldNotValidException("상품의 금액은 최소 10원, 최대 1억원까지 입력할 수 있습니다.");
         }
     }
