+++ src/test/java/cart/JwpCartApplicationTests.java
@@ -110,7 +110,7 @@ class JwpCartApplicationTests {
                     .put("/items/{id}", 1L)
                     .then().log().all()
                     .contentType(ContentType.JSON)
-                    .statusCode(HttpStatus.BAD_REQUEST.value())
+                    .statusCode(HttpStatus.NOT_FOUND.value())
                     .body("message", is("일치하는 상품을 찾을 수 없습니다."));
         }
 
@@ -124,7 +124,7 @@ class JwpCartApplicationTests {
                     .delete("/items/{id}", 1L)
                     .then().log().all()
                     .contentType(ContentType.JSON)
-                    .statusCode(HttpStatus.BAD_REQUEST.value())
+                    .statusCode(HttpStatus.NOT_FOUND.value())
                     .body("message", is("일치하는 상품을 찾을 수 없습니다."));
         }
     }
